requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
psycopg2==2.9.9